requirements.txt
absl-py==0.15.0
astunparse==1.6.3
backcall==0.2.0
black==21.11b1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==7.1.2
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
flatbuffers==1.12
fonttools==4.28.2
gast==0.4.0
google-api-core==1.31.4
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
Google-Images-Search==1.3.9
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.20.2
idna==3.3
imageio==2.12.0
importlib-metadata==4.8.2
ipykernel==6.5.1
ipython==7.29.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lime==0.2.0.1
Markdown==3.3.6
matplotlib==3.5.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.1
networkx==2.6.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
Pygments==2.10.0
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
python-resize-image==1.1.20
pytz==2021.3
PyWavelets==1.2.0
pywin32==302
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.15.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
tweepy==3.8.0
typing_extensions==4.0.0
uritemplate==3.0.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
windows-curses==2.3.0
wrapt==1.12.1
zipp==3.6.0
